# apidocs/openapi.yaml
openapi: 3.0.3
info:
  title: Atom
  version: 0.1.0
  description: |
    Identity and Authorization service for IoT and cloud-native systems.

    All endpoints except `GET /health`, `POST /auth/login`, and
    `GET /.well-known/jwks.json` require a Bearer token:

    ```
    Authorization: Bearer <token>
    ```

    Two token types are accepted:
    - **JWT** (`eyJ...`) — returned by `/auth/login`, short-lived (default 1 h)
    - **API key** (`atom_...`) — long-lived, created per entity

    Endpoints marked **RequireManage** additionally require the caller to hold an
    `allow` + `scope=all` policy binding covering the `manage` capability.
  license:
    name: Apache-2.0
servers:
  - url: http://localhost:8080
    description: Local development
security:
  - BearerAuth: []

# ---------------------------------------------------------------------------
# Reusable components
# ---------------------------------------------------------------------------
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: JWT (`eyJ...`) or API key (`atom_<id>_<secret>`)

  # ── Schemas ──────────────────────────────────────────────────────────────
  schemas:
...
    EntityKind:
      type: string
      enum: [human, device, service, workload, application]
    EntityStatus:
      type: string
      enum: [active, inactive, suspended]
      default: active
    ProfileStatus:
      type: string
      enum: [active, deprecated, disabled]
      default: active
    ProfileVersionStatus:
      type: string
      enum: [draft, active, deprecated, disabled]
      default: active
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: not found

    # ── Auth ───────────────────────────────────────────────────────────────
...
    Entity:
      type: object
      required: [id, kind, name, status, attributes, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        kind:
          $ref: '#/components/schemas/EntityKind'
        name:
          type: string
        tenant_id:
          type: string
          format: uuid
          nullable: true
        profile_id:
          type: string
          format: uuid
          nullable: true
          description: User/domain profile. Authorization still uses `kind`.
        profile_version_id:
          type: string
          format: uuid
          nullable: true
          description: Profile version used to validate `attributes`.
        status:
          $ref: '#/components/schemas/EntityStatus'
        attributes:
          type: object
          description: Arbitrary JSONB metadata, usable in ABAC conditions
          example: {"department": "engineering", "env": "prod"}
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateEntity:
      type: object
      required: [name]
      description: >
        Provide either `kind` for legacy/internal creation, or `profile_id` to derive
        the internal Atom kind from an active entity profile. `profile_version_id`
        is optional when `profile_id` is present; Atom uses the active version
        when omitted.
      properties:
        kind:
          $ref: '#/components/schemas/EntityKind'
        profile_id:
          type: string
          format: uuid
        profile_version_id:
          type: string
          format: uuid
        name:
          type: string
          example: sensor-01
        tenant_id:
          type: string
          format: uuid
        attributes:
          type: object
    UpdateEntity:
      type: object
      properties:
        name:
          type: string
        status:
          $ref: '#/components/schemas/EntityStatus'
        attributes:
          type: object
          description: Replaces all existing attributes when provided
    EntityList:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Entity'
        total:
          type: integer
          description: Total count across all pages

    # ── Profiles ───────────────────────────────────────────────────────────
    Profile:
      type: object
      required: [id, object_kind, kind, key, display_name, status, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
          nullable: true
        object_kind:
          type: string
          enum: [entity, resource, group, tenant, credential]
        kind:
          type: string
          description: Internal object kind value. Entity profiles must use an EntityKind.
          example: device
        key:
          type: string
          description: User/domain subtype key.
          example: client
        display_name:
          type: string
          example: Client
        description:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/ProfileStatus'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateProfile:
      type: object
      required: [object_kind, kind, key, display_name]
      properties:
        tenant_id:
          type: string
          format: uuid
        object_kind:
          type: string
          enum: [entity, resource, group, tenant, credential]
        kind:
          type: string
          example: device
        key:
          type: string
          example: client
        display_name:
          type: string
          example: Client
        description:
          type: string
        status:
          $ref: '#/components/schemas/ProfileStatus'
    ProfileVersion:
      type: object
      required: [id, profile_id, version, json_schema, ui_schema, status, created_at]
      properties:
        id:
          type: string
          format: uuid
        profile_id:
          type: string
          format: uuid
        version:
          type: integer
        json_schema:
          type: object
          default: {}
        ui_schema:
          type: object
          default: {}
        status:
          $ref: '#/components/schemas/ProfileVersionStatus'
        created_at:
          type: string
          format: date-time
    CreateProfileVersion:
      type: object
      required: [version]
      properties:
        version:
          type: integer
        json_schema:
          type: object
          default: {}
        ui_schema:
          type: object
          default: {}
        status:
          $ref: '#/components/schemas/ProfileVersionStatus'
    ProfileList:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Profile'
        total:
          type: integer
    ProfileVersionList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ProfileVersion'

    # ── Credentials ────────────────────────────────────────────────────────
...
    CreateApiKey:
      type: object
      properties:
        expires_at:
          type: string
          format: date-time
          description: Omit for no expiry
        description:
          type: string
          example: device-01 production key
    ApiKeyResponse:
      type: object
      required: [credential_id, key]
      properties:
        credential_id:
          type: string
          format: uuid
        key:
          type: string
          description: >
            Full API key — format `atom_<32-hex-id>_<64-hex-secret>`.
            Shown exactly once. Store securely.
          example: atom_0a1b2c3d...
        expires_at:
          type: string
          format: date-time
          nullable: true

    # ── Groups ─────────────────────────────────────────────────────────────
    Group:
      type: object
      required: [id, name, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        tenant_id:
          type: string
          format: uuid
          nullable: true
        description:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateGroup:
      type: object
      required: [name]
      properties:
        name:
          type: string
          example: floor-sensors
        tenant_id:
          type: string
          format: uuid
        description:
          type: string
    GroupList:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Group'
        total:
          type: integer
...
  responses:
    Unauthorized:
      description: Missing or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: missing Authorization header
    Forbidden:
      description: Caller lacks `manage` scope=all
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: forbidden
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: not found
    Conflict:
      description: Already exists (unique constraint)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: already exists

# ---------------------------------------------------------------------------
# Paths
# ---------------------------------------------------------------------------
paths:
...
  /entities:
    get:
      operationId: listEntities
      summary: List entities
      tags: [Entities]
      parameters:
        - name: kind
          in: query
          schema:
            $ref: '#/components/schemas/EntityKind'
        - name: profile_id
          in: query
          schema:
            type: string
            format: uuid
        - name: tenant_id
          in: query
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/EntityStatus'
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Paginated entity list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EntityList'
        "401":
          $ref: '#/components/responses/Unauthorized'

    post:
      operationId: createEntity
      summary: Create an entity
      tags: [Entities]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEntity'
            example:
              kind: device
              profile_id: null
              name: sensor-01
              attributes:
                firmware: "2.1.0"
      responses:
        "201":
          description: Entity created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Entity'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "409":
          $ref: '#/components/responses/Conflict'
  /entities/{id}:
    get:
      operationId: getEntity
      summary: Get an entity
      tags: [Entities]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Entity'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "404":
          $ref: '#/components/responses/NotFound'

    put:
      operationId: updateEntity
      summary: Update an entity
      description: All fields are optional; only provided fields are updated (COALESCE semantics).
      tags: [Entities]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEntity'
      responses:
        "200":
          description: Updated entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Entity'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "404":
          $ref: '#/components/responses/NotFound'

    delete:
      operationId: deleteEntity
      summary: Delete an entity
      description: >
        An entity may always delete itself.
        Deleting any other entity requires `manage` scope=all (**RequireManage**).
      tags: [Entities]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Deleted
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'

  # ── Profiles ───────────────────────────────────────────────────────────
  /profiles:
    get:
      operationId: listProfiles
      summary: List profiles
      tags: [Profiles]
      parameters:
        - name: object_kind
          in: query
          schema:
            type: string
            enum: [entity, resource, group, tenant, credential]
        - name: kind
          in: query
          schema:
            type: string
        - name: key
          in: query
          schema:
            type: string
        - name: tenant_id
          in: query
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/ProfileStatus'
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Paginated profile list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileList'
        "401":
          $ref: '#/components/responses/Unauthorized'

    post:
      operationId: createProfile
      summary: Create a profile
      description: "**RequireManage.** Entity profiles derive `entities.kind`; authorization still uses that internal kind."
      tags: [Profiles]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProfile'
            example:
              object_kind: entity
              kind: device
              key: client
              display_name: Client
      responses:
        "201":
          description: Profile created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "409":
          $ref: '#/components/responses/Conflict'
...
  /profiles/{id}/versions:
    get:
      operationId: listProfileVersions
      summary: List profile versions
      tags: [Profiles]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Profile versions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileVersionList'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "404":
          $ref: '#/components/responses/NotFound'

    post:
      operationId: createProfileVersion
      summary: Create a profile version
      description: "**RequireManage.** Creating a version does not mutate older versions."
      tags: [Profiles]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProfileVersion'
            example:
              version: 1
              json_schema: {}
              ui_schema: {}
      responses:
        "201":
          description: Profile version created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileVersion'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          $ref: '#/components/responses/NotFound'
        "409":
          $ref: '#/components/responses/Conflict'

  # ── Credentials ────────────────────────────────────────────────────────
...
  /entities/{id}/credentials/api-keys:
    post:
      operationId: createApiKey
      summary: Create an API key
      description: >
        The full key (`atom_<id>_<secret>`) is returned exactly once and never
        stored in plaintext. Store it securely immediately.
      tags: [Credentials]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateApiKey'
      responses:
        "201":
          description: API key created — store the `key` value immediately
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKeyResponse'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "404":
          $ref: '#/components/responses/NotFound'
...
  /groups:
    get:
      operationId: listGroups
      summary: List groups
      tags: [Groups]
      parameters:
        - name: tenant_id
          in: query
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Group list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupList'
        "401":
          $ref: '#/components/responses/Unauthorized'

    post:
      operationId: createGroup
      summary: Create a group
      tags: [Groups]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGroup'
      responses:
        "201":
          description: Group created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "409":
          $ref: '#/components/responses/Conflict'
...